+++ activesupport/lib/active_support/core_ext/string/starts_ends_with.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 class String
-  alias_method :starts_with?, :start_with?
-  alias_method :ends_with?, :end_with?
-  deprecate starts_with?: :start_with?
-  deprecate ends_with?: :end_with?
+  alias :starts_with? :start_with?
+  alias :ends_with? :end_with?
 end